refactor(codegen): extrai compile/{user_fn,main_fn}.rs de func.rs
- compile/user_fn.rs (~298 LOC) — compile_user_fn: orquestra
  HIR→MIR→Cranelift com fallback automatico para AST quando MIR
  bail. Importa try_compile_via_mir do compile::mir_route.
- compile/main_fn.rs (~178 LOC) — compile_main + helper sintetico
  compile_main_entry_shim (entry C-style chamando __RTS_MAIN).

Helpers user_call_conv e collect_var_decls em func.rs viraram
pub(crate) para serem chamados dos novos modulos.

func.rs: 1167 -> 686 linhas. Suite verde (11/11 crates).
